moved test into separate file
